+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Version 0.7.0
+
+0.7.0 plan: <https://github.com/greyli/apiflask/issues/46>
+
 ## Version 0.6.0
 
 Released: 2021/5/11